+++ src/Smilei.cpp
@@ -154,7 +154,7 @@ int main( int argc, char *argv[] )
             region.define_regions_map(target_map, &smpi, params);
 
             // params.map_rank used to defined regions neighborood
-            region.build( params, &smpi, vecPatches, openPMD, false );
+            region.build( params, &smpi, vecPatches, openPMD, false, simWindow->getNmoved() );
             region.identify_additional_patches( &smpi, vecPatches, params, simWindow );
             region.identify_missing_patches( &smpi, vecPatches, params );
         }
@@ -205,7 +205,7 @@ int main( int argc, char *argv[] )
         if( params.multiple_decomposition ) {
             TITLE( "Create SDMD grids" );
             region.vecPatch_.refHindex_ = smpi.getRank();
-            region.build( params, &smpi, vecPatches, openPMD, false );
+            region.build( params, &smpi, vecPatches, openPMD, false, simWindow->getNmoved() );
             region.identify_additional_patches( &smpi, vecPatches, params, simWindow );
             region.identify_missing_patches( &smpi, vecPatches, params );
             //cout << smpi.getRank() << "\t - local : " << region.local_patches_.size()
@@ -216,7 +216,7 @@ int main( int argc, char *argv[] )
             region.clean();
             region.reset_mapping();
             
-            region.build( params, &smpi, vecPatches, openPMD, false );
+            region.build( params, &smpi, vecPatches, openPMD, false, simWindow->getNmoved() );
             region.identify_additional_patches( &smpi, vecPatches, params, simWindow );
             region.identify_missing_patches( &smpi, vecPatches, params );
             //cout << smpi.getRank() << "\t - local : " << region.local_patches_.size()
@@ -282,7 +282,7 @@ int main( int argc, char *argv[] )
         if( params.initial_rotational_cleaning ) {
             TITLE( "Rotational cleaning" );
             Region region_global( params );
-            region_global.build( params, &smpi, vecPatches, openPMD, true );
+            region_global.build( params, &smpi, vecPatches, openPMD, true, simWindow->getNmoved() );
             region_global.identify_additional_patches( &smpi, vecPatches, params, simWindow );
             region_global.identify_missing_patches( &smpi, vecPatches, params );
             for (unsigned int imode = 0 ; imode < params.nmodes ; imode++  ) {
@@ -588,7 +588,7 @@ int main( int argc, char *argv[] )
                     region.reset_fitting( &smpi, params );
                     region.clean();
                     region.reset_mapping();
-                    region.build( params, &smpi, vecPatches, openPMD, false );
+                    region.build( params, &smpi, vecPatches, openPMD, false, simWindow->getNmoved() );
                     if( params.is_pxr ) {
                         region.coupling( params, false );
                     }
@@ -692,7 +692,7 @@ int executeTestMode( VectorPatch &vecPatches,
     if( params.restart ) {
         if (params.multiple_decomposition) {
             checkpoint.readRegionDistribution( region );
-            region.build( params, smpi, vecPatches, openPMD, false );
+            region.build( params, smpi, vecPatches, openPMD, false, simWindow->getNmoved() );
         }
         checkpoint.restartAll( vecPatches, region, smpi, simWindow, params, openPMD );
     }